# feat: optimize graph based on sync.Map