--- a/pkg/graph/dag/dag_test.go
+++ b/pkg/graph/dag/dag_test.go
@@ -179,17 +179,17 @@ func TestDAG_VertexCount(t *testing.T) {
 				}
 
 				d.VertexCount()
-				assert.Equal(d.VertexCount(), 1)
+				assert.Equal(d.VertexCount(), uint64(1))
 
 				d.DeleteVertex(mockVertexID)
-				assert.Equal(d.VertexCount(), 0)
+				assert.Equal(d.VertexCount(), uint64(0))
 			},
 		},
 		{
 			name: "empty dag",
 			expect: func(t *testing.T, d DAG[string]) {
 				assert := assert.New(t)
-				assert.Equal(d.VertexCount(), 0)
+				assert.Equal(d.VertexCount(), uint64(0))
 			},
 		},
 	}
@@ -290,38 +290,6 @@ func TestDAG_GetRandomVertices(t *testing.T) {
 
 				vertices = d.GetRandomVertices(1)
 				assert.Equal(len(vertices), 0)
-			},
-		},
-	}
-
-	for _, tc := range tests {
-		t.Run(tc.name, func(t *testing.T) {
-			d := NewDAG[string]()
-			tc.expect(t, d)
-		})
-	}
-}
-
-func TestDAG_GetVertexKeys(t *testing.T) {
-	tests := []struct {
-		name   string
-		expect func(t *testing.T, d DAG[string])
-	}{
-		{
-			name: "get keys of vertices",
-			expect: func(t *testing.T, d DAG[string]) {
-				assert := assert.New(t)
-				if err := d.AddVertex(mockVertexID, mockVertexValue); err != nil {
-					assert.NoError(err)
-				}
-
-				keys := d.GetVertexKeys()
-				assert.Equal(len(keys), 1)
-				assert.Equal(keys[0], mockVertexID)
-
-				d.DeleteVertex(mockVertexID)
-				keys = d.GetVertexKeys()
-				assert.Equal(len(keys), 0)
 			},
 		},
 	}
@@ -965,7 +933,25 @@ func BenchmarkDAG_DeleteVertex(b *testing.B) {
 	}
 }
 
-func BenchmarkDAG_GetRandomKeys(b *testing.B) {
+func BenchmarkDAG_GetVertices(b *testing.B) {
+	d := NewDAG[string]()
+	for n := 0; n < b.N; n++ {
+		id := fmt.Sprint(n)
+		if err := d.AddVertex(id, string(id)); err != nil {
+			b.Fatal(err)
+		}
+	}
+
+	b.ResetTimer()
+	for n := 0; n < b.N; n++ {
+		vertices := d.GetVertices()
+		if len(vertices) != b.N {
+			b.Fatal(errors.New("get vertices failed"))
+		}
+	}
+}
+
+func BenchmarkDAG_GetRandomVertices(b *testing.B) {
 	d := NewDAG[string]()
 	for n := 0; n < b.N; n++ {
 		id := fmt.Sprint(n)