--- a/pkg/graph/dg/dg.go
+++ b/pkg/graph/dg/dg.go
@@ -20,11 +20,9 @@ package dg
 
 import (
 	"errors"
-	"math/rand"
 	"sync"
-	"time"
 
-	cmap "github.com/orcaman/concurrent-map/v2"
+	"go.uber.org/atomic"
 
 	"d7y.io/dragonfly/v2/pkg/container/set"
 )
@@ -32,6 +30,9 @@ import (
 var (
 	// ErrVertexNotFound represents vertex not found.
 	ErrVertexNotFound = errors.New("vertex not found")
+
+	// ErrVertexInvalid represents vertex invalid.
+	ErrVertexInvalid = errors.New("vertex invalid")
 
 	// ErrVertexAlreadyExists represents vertex already exists.
 	ErrVertexAlreadyExists = errors.New("vertex already exists")
@@ -63,9 +64,6 @@ type DG[T comparable] interface {
 	// GetRandomVertices returns random map of vertices.
 	GetRandomVertices(n uint) []*Vertex[T]
 
-	// GetVertexKeys returns keys of vertices.
-	GetVertexKeys() []string
-
 	// GetSourceVertices returns source vertices.
 	GetSourceVertices() []*Vertex[T]
 
@@ -73,7 +71,7 @@ type DG[T comparable] interface {
 	GetSinkVertices() []*Vertex[T]
 
 	// VertexCount returns count of vertices.
-	VertexCount() int
+	VertexCount() uint64
 
 	// AddEdge adds edge between two vertices.
 	AddEdge(fromVertexID, toVertexID string) error
@@ -81,7 +79,7 @@ type DG[T comparable] interface {
 	// DeleteEdge deletes edge between two vertices.
 	DeleteEdge(fromVertexID, toVertexID string) error
 
-	// CanAddEdge finds whether there are circles through depth-first search.
+	// CanAddEdge indicates whether can add edge between two vertices.
 	CanAddEdge(fromVertexID, toVertexID string) bool
 
 	// DeleteVertexInEdges deletes inedges of vertex.
@@ -93,14 +91,17 @@ type DG[T comparable] interface {
 
 // dg provides directed graph function.
 type dg[T comparable] struct {
+	vertices *sync.Map
+	count    *atomic.Uint64
 	mu       sync.RWMutex
-	vertices cmap.ConcurrentMap[string, *Vertex[T]]
 }
 
 // New returns a new DG interface.
 func NewDG[T comparable]() DG[T] {
 	return &dg[T]{
-		vertices: cmap.New[*Vertex[T]](),
+		vertices: &sync.Map{},
+		count:    atomic.NewUint64(0),
+		mu:       sync.RWMutex{},
 	}
 }
 
@@ -109,11 +110,11 @@ func (d *dg[T]) AddVertex(id string, value T) error {
 	d.mu.Lock()
 	defer d.mu.Unlock()
 
-	if _, ok := d.vertices.Get(id); ok {
+	if _, loaded := d.vertices.LoadOrStore(id, NewVertex(id, value)); loaded {
 		return ErrVertexAlreadyExists
 	}
 
-	d.vertices.Set(id, NewVertex(id, value))
+	d.count.Inc()
 	return nil
 }
 
@@ -122,7 +123,12 @@ func (d *dg[T]) DeleteVertex(id string) {
 	d.mu.Lock()
 	defer d.mu.Unlock()
 
-	vertex, ok := d.vertices.Get(id)
+	rawVertex, loaded := d.vertices.Load(id)
+	if !loaded {
+		return
+	}
+
+	vertex, ok := rawVertex.(*Vertex[T])
 	if !ok {
 		return
 	}
@@ -136,14 +142,20 @@ func (d *dg[T]) DeleteVertex(id string) {
 		continue
 	}
 
-	d.vertices.Remove(id)
+	d.vertices.Delete(id)
+	d.count.Dec()
 }
 
 // GetVertex gets vertex from graph.
 func (d *dg[T]) GetVertex(id string) (*Vertex[T], error) {
-	vertex, ok := d.vertices.Get(id)
-	if !ok {
+	rawVertex, loaded := d.vertices.Load(id)
+	if !loaded {
 		return nil, ErrVertexNotFound
+	}
+
+	vertex, ok := rawVertex.(*Vertex[T])
+	if !ok {
+		return nil, ErrVertexInvalid
 	}
 
 	return vertex, nil
@@ -151,7 +163,26 @@ func (d *dg[T]) GetVertex(id string) (*Vertex[T], error) {
 
 // GetVertices returns map of vertices.
 func (d *dg[T]) GetVertices() map[string]*Vertex[T] {
-	return d.vertices.Items()
+	d.mu.RLock()
+	defer d.mu.RUnlock()
+
+	vertices := make(map[string]*Vertex[T], d.count.Load())
+	d.vertices.Range(func(key, value interface{}) bool {
+		vertex, ok := value.(*Vertex[T])
+		if !ok {
+			return true
+		}
+
+		id, ok := key.(string)
+		if !ok {
+			return true
+		}
+
+		vertices[id] = vertex
+		return true
+	})
+
+	return vertices
 }
 
 // GetRandomVertices returns random map of vertices.
@@ -159,32 +190,27 @@ func (d *dg[T]) GetRandomVertices(n uint) []*Vertex[T] {
 	d.mu.RLock()
 	defer d.mu.RUnlock()
 
-	keys := d.GetVertexKeys()
-	if int(n) >= len(keys) {
-		n = uint(len(keys))
+	if n == 0 {
+		return nil
 	}
 
-	r := rand.New(rand.NewSource(time.Now().UnixNano()))
-	permutation := r.Perm(len(keys))[:n]
 	randomVertices := make([]*Vertex[T], 0, n)
-	for _, v := range permutation {
-		key := keys[v]
-		if vertex, err := d.GetVertex(key); err == nil {
-			randomVertices = append(randomVertices, vertex)
+	d.vertices.Range(func(key, value interface{}) bool {
+		vertex, ok := value.(*Vertex[T])
+		if !ok {
+			return true
 		}
-	}
+
+		randomVertices = append(randomVertices, vertex)
+		return uint(len(randomVertices)) < n
+	})
 
 	return randomVertices
 }
 
-// GetVertexKeys returns keys of vertices.
-func (d *dg[T]) GetVertexKeys() []string {
-	return d.vertices.Keys()
-}
-
 // VertexCount returns count of vertices.
-func (d *dg[T]) VertexCount() int {
-	return d.vertices.Count()
+func (d *dg[T]) VertexCount() uint64 {
+	return d.count.Load()
 }
 
 // AddEdge adds edge between two vertices.
@@ -196,14 +222,14 @@ func (d *dg[T]) AddEdge(fromVertexID, toVertexID string) error {
 		return ErrCycleBetweenVertices
 	}
 
-	fromVertex, ok := d.vertices.Get(fromVertexID)
-	if !ok {
-		return ErrVertexNotFound
+	fromVertex, err := d.GetVertex(fromVertexID)
+	if err != nil {
+		return err
 	}
 
-	toVertex, ok := d.vertices.Get(toVertexID)
-	if !ok {
-		return ErrVertexNotFound
+	toVertex, err := d.GetVertex(toVertexID)
+	if err != nil {
+		return err
 	}
 
 	for _, child := range fromVertex.Children.Values() {
@@ -228,14 +254,14 @@ func (d *dg[T]) DeleteEdge(fromVertexID, toVertexID string) error {
 	d.mu.Lock()
 	defer d.mu.Unlock()
 
-	fromVertex, ok := d.vertices.Get(fromVertexID)
-	if !ok {
-		return ErrVertexNotFound
+	fromVertex, err := d.GetVertex(fromVertexID)
+	if err != nil {
+		return err
 	}
 
-	toVertex, ok := d.vertices.Get(toVertexID)
-	if !ok {
-		return ErrVertexNotFound
+	toVertex, err := d.GetVertex(toVertexID)
+	if err != nil {
+		return err
 	}
 
 	fromVertex.Children.Delete(toVertex)
@@ -243,7 +269,7 @@ func (d *dg[T]) DeleteEdge(fromVertexID, toVertexID string) error {
 	return nil
 }
 
-// CanAddEdge finds whether there are circles through depth-first search.
+// CanAddEdge indicates whether can add edge between two vertices.
 func (d *dg[T]) CanAddEdge(fromVertexID, toVertexID string) bool {
 	d.mu.RLock()
 	defer d.mu.RUnlock()
@@ -252,12 +278,12 @@ func (d *dg[T]) CanAddEdge(fromVertexID, toVertexID string) bool {
 		return false
 	}
 
-	fromVertex, ok := d.vertices.Get(fromVertexID)
-	if !ok {
+	fromVertex, err := d.GetVertex(fromVertexID)
+	if err != nil {
 		return false
 	}
 
-	if _, ok := d.vertices.Get(toVertexID); !ok {
+	if _, err := d.GetVertex(toVertexID); err != nil {
 		return false
 	}
 
@@ -275,9 +301,9 @@ func (d *dg[T]) DeleteVertexInEdges(id string) error {
 	d.mu.Lock()
 	defer d.mu.Unlock()
 
-	vertex, ok := d.vertices.Get(id)
-	if !ok {
-		return ErrVertexNotFound
+	vertex, err := d.GetVertex(id)
+	if err != nil {
+		return err
 	}
 
 	for _, parent := range vertex.Parents.Values() {
@@ -293,9 +319,9 @@ func (d *dg[T]) DeleteVertexOutEdges(id string) error {
 	d.mu.Lock()
 	defer d.mu.Unlock()
 
-	vertex, ok := d.vertices.Get(id)
-	if !ok {
-		return ErrVertexNotFound
+	vertex, err := d.GetVertex(id)
+	if err != nil {
+		return err
 	}
 
 	for _, child := range vertex.Children.Values() {
@@ -312,7 +338,7 @@ func (d *dg[T]) GetSourceVertices() []*Vertex[T] {
 	defer d.mu.RUnlock()
 
 	var sourceVertices []*Vertex[T]
-	for _, vertex := range d.vertices.Items() {
+	for _, vertex := range d.GetVertices() {
 		if vertex.InDegree() == 0 {
 			sourceVertices = append(sourceVertices, vertex)
 		}
@@ -327,26 +353,11 @@ func (d *dg[T]) GetSinkVertices() []*Vertex[T] {
 	defer d.mu.RUnlock()
 
 	var sinkVertices []*Vertex[T]
-	for _, vertex := range d.vertices.Items() {
+	for _, vertex := range d.GetVertices() {
 		if vertex.OutDegree() == 0 {
 			sinkVertices = append(sinkVertices, vertex)
 		}
 	}
 
 	return sinkVertices
-}
-
-// search finds successors of vertex.
-func (d *dg[T]) search(vertexID string, successors map[string]struct{}) {
-	vertex, ok := d.vertices.Get(vertexID)
-	if !ok {
-		return
-	}
-
-	for _, child := range vertex.Children.Values() {
-		if _, ok := successors[child.ID]; !ok {
-			successors[child.ID] = struct{}{}
-			d.search(child.ID, successors)
-		}
-	}
 }